+++ core/src/main/php/peer/http/BasicAuthorization.class.php
@@ -38,7 +38,7 @@
     }
 
     /** @return string */
-    public function getUser() { return $this->user; }
+    public function getUser() { return $this->username; }
     
     /**
      * Returns a BasicAuthorization object from header value; returns
@@ -84,6 +84,15 @@
         'Authorization',
         new Header('Authorization', $this->getValueRepresentation())
       );
+    }
+
+    /**
+     * Retrieve string representation
+     *
+     * @return string
+     */
+    public function toString() {
+      return $this->getClassName().' { username= "'.$this->username.'" }';
     }
   }
 ?>