+++ core/src/main/php/webservices/rest/ResponseReader.class.php
@@ -27,14 +27,15 @@
      *
      * @param  lang.Type $t
      * @param  io.streams.InputStream $is
+     * @param   string $encoding
      * @return var
      */
-    public function read(Type $t, InputStream $is) {
+    public function read(Type $t, InputStream $is, $encoding= xp::ENCODING) {
       if (NULL === $this->deserializer) {
         throw new IllegalStateException('No deserializer available.');
       }
 
-      return $this->marshalling->unmarshal($t, $this->deserializer->deserialize($is));
+      return $this->marshalling->unmarshal($t, $this->deserializer->deserialize($is, $encoding));
     }
   }
 ?>